+++ src/ZPublisher/tests/test_utils.py
@@ -36,6 +36,55 @@ class SafeUnicodeTests(unittest.TestCase):
         self.assertEqual(self._makeOne(b'test\xc2\xae'), 'test\xae')
 
 
+class NoUpdatePropertyManager:
+    """PropertyManager without _updateProperty method.
+
+    This is a simplified version of the original PropertyManager,
+    with only the methods we need.
+    """
+    _properties = ()
+
+    def _setPropValue(self, id, value):
+        if type(value) == list:
+            value = tuple(value)
+        setattr(self, id, value)
+
+    def _setProperty(self, id, value, type='string'):
+        self._properties = self._properties + ({'id': id, 'type': type},)
+        self._setPropValue(id, value)
+
+    def hasProperty(self, id):
+        for p in self._properties:
+            if id == p['id']:
+                return 1
+        return 0
+
+    def getProperty(self, id, d=None):
+        if self.hasProperty(id):
+            return getattr(self, id)
+        return d
+
+    def getPropertyType(self, id):
+        for md in self._properties:
+            if md['id'] == id:
+                return md.get('type', 'string')
+        return None
+
+    def _propertyMap(self):
+        return self._properties
+
+    def propertyMap(self):
+        return tuple(dict.copy() for dict in self._propertyMap())
+
+
+class NoPropertiesManager(NoUpdatePropertyManager):
+    """PropertyManager with _updateProperty method but without _properties."""
+    _properties = None
+
+    def _updateProperty(self, id, value):
+        self._setPropValue(id, value)
+
+
 class FixPropertiesTests(unittest.TestCase):
 
     def _makeOne(self):
@@ -102,3 +151,28 @@ class FixPropertiesTests(unittest.TestCase):
         fix_properties(obj)
         self.assertEqual(obj.getProperty("prop1"), "single line")
         self.assertEqual(obj.getPropertyType("prop1"), "string")
+
+    def test_no_update(self):
+        # Test that an object without _updateProperty method does not trip up
+        # our code.
+        from ZPublisher.utils import fix_properties
+
+        obj = NoUpdatePropertyManager()
+        obj._setProperty("mixed", ["text and", b"bytes"], "lines")
+        self.assertEqual(obj.getProperty("mixed"), ("text and", b"bytes"))
+        self.assertEqual(obj.getPropertyType("mixed"), "lines")
+
+        # This should not raise an error.
+        fix_properties(obj)
+        # The properties should have remained the same.
+        self.assertEqual(obj.getProperty("mixed"), ("text and", b"bytes"))
+        self.assertEqual(obj.getPropertyType("mixed"), "lines")
+
+    def test_no_properties(self):
+        # Test that an object with a failing propertyMap method,
+        # due to _properties=None, does not trip up our code.
+        from ZPublisher.utils import fix_properties
+
+        obj = NoPropertiesManager()
+        # This should not raise an error.
+        fix_properties(obj)