+++ src/ZPublisher/utils.py
@@ -145,11 +145,19 @@ def fix_properties(obj, path=None):
             # We don't care about the path then, it is only shown in logs.
             path = "/dummy"
 
+    if not hasattr(obj, "_updateProperty"):
+        # Seen with portal_url tool, most items in portal_skins,
+        # catalog lexicons, workflow states/transitions/variables, etc.
+        return
     try:
         prop_map = obj.propertyMap()
-    except AttributeError:
-        # Object does not inherit from PropertyManager.
-        # For example 'MountedObject'.
+    except (AttributeError, TypeError, KeyError, ValueError):
+        # If getting the property map fails, there is nothing we can do.
+        # Problems seen in practice:
+        # - Object does not inherit from PropertyManager,
+        #   for example 'MountedObject'.
+        # - Object is a no longer existing skin layer.
+        logger.warning("Error getting property map from %s", path)
         return
 
     for prop_info in prop_map: